requirements.txt
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
googleapis-common-protos==1.73.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
linkify-it-py==2.1.0
markdown-it-py==4.0.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
oauthlib==3.3.1
platformdirs==4.9.4
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.19.2
pyparsing==3.3.2
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.3
rsa==4.9.1
textual==8.1.1
typing_extensions==4.15.0
uc-micro-py==2.0.0
uritemplate==4.2.0
urllib3==2.6.3
